rearrange the directory structure of the project
