+++ admin-app/frontend/src/components/auth/ForgotPasswordForm.vue
