+++ admin-app/frontend/src/components/auth/GoogleIcon.vue
