+++ admin-app/frontend/src/components/commons/DataCreationButtonCard.vue
@@ -13,7 +13,7 @@
 
 <script>
 import { mapActions } from 'vuex';
-import { dataTypesDict } from '../../store/constants';
+import { dataTypesDict } from '../../lib/constants';
 
 export default {
   name: 'DataCreationButtonCard',
