+++ admin-app/frontend/src/components/commons/LocaleTextBox.vue
@@ -38,9 +38,9 @@
 
 <script>
 import { mapGetters } from 'vuex';
-import { mutableI18nString } from '../../store/classes';
+import { mutableI18nString } from '../../lib/classes';
 // import { I18nString } from '../../models';
-// import { emptyI18nString } from '../../store/classes';
+// import { emptyI18nString } from '../../lib/classes';
 
 export default {
   props: {
