+++ admin-app/frontend/src/components/interventions/Intervention.vue
@@ -21,8 +21,8 @@
 <script>
 // import { validate as uuidValidate } from 'uuid';
 import { mapGetters, mapActions } from 'vuex';
-import { dataTypesDict } from '../../store/constants';
-import { deriveFilePath } from '../../store/utils';
+import { dataTypesDict } from '../../lib/constants';
+import { deriveFilePath } from '../../lib/utils';
 import ImgFromS3 from '../commons/ImgFromS3.vue';
 
 export default {
