+++ admin-app/frontend/src/components/surveys/Survey.vue
@@ -30,8 +30,8 @@
 
 <script>
 import { mapGetters, mapActions } from 'vuex';
-import { dataTypesDict } from '../../store/constants';
-import { deriveFilePath } from '../../store/utils';
+import { dataTypesDict } from '../../lib/constants';
+import { deriveFilePath } from '../../lib/utils';
 import ImgFromS3 from '../commons/ImgFromS3.vue';
 
 export default {
