+++ admin-app/frontend/src/components/surveys/surveyModalContent/SurveyModalQuestion.vue
@@ -355,12 +355,12 @@ import {
   emptyMutableQuestionOption,
   mutableI18nString,
   mutableQuestionOption,
-} from '../../../store/classes';
-import { modalModesDict, questionTypesIconDict } from '../../../store/constants';
+} from '../../../lib/classes';
+import { modalModesDict, questionTypesIconDict } from '../../../lib/constants';
 // eslint-disable-next-line import/named
-import { compareI18nStrings, deriveFilePath } from '../../../store/utils';
+import { compareI18nStrings, deriveFilePath } from '../../../lib/utils';
 
-import LocaleTextBox from '../../global/LocaleTextBox.vue';
+import LocaleTextBox from '../../commons/LocaleTextBox.vue';
 import FileInput from '../../commons/FileInput.vue';
 import ImgFromS3 from '../../commons/ImgFromS3.vue';
