+++ admin-app/frontend/src/i18n/en-US.json
