+++ admin-app/frontend/src/i18n/es-BO.json
