+++ admin-app/frontend/src/lib/classes.js
