+++ admin-app/frontend/src/store/entities/data.js
@@ -1,6 +1,6 @@
 import { DataStore } from '@aws-amplify/datastore';
 import { API } from 'aws-amplify';
-import { dataTypesDict, modalModesDict } from '../constants';
+import { dataTypesDict, modalModesDict } from '../../lib/constants';
 import { deleteEntity } from '../../graphql/mutations';
 import { Entity, I18nString } from '../../models';
