+++ admin-app/frontend/src/views/OrganizationStructure.vue
@@ -41,7 +41,7 @@ import LevelModal from '../components/organizationStructure/LevelModal.vue';
 import EntityModal from '../components/organizationStructure/EntityModal.vue';
 import EntitiesColumn from '../components/organizationStructure/EntitiesColumn.vue';
 import LevelColumnHeader from '../components/organizationStructure/LevelColumnHeader.vue';
-import { dataTypesDict } from '../store/constants';
+import { dataTypesDict } from '../lib/constants';
 
 export default {
   name: 'OrganizationStructure',
